% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy.R
\name{AddNode}
\alias{AddNode}
\title{Add node to a taxonomic tree}
\usage{
AddNode(tax, name, parent = c(), distance = 1, count = NA)
}
\arguments{
\item{tax}{object of class \code{Taxonomy}}

\item{name}{string: name of new node}

\item{parent}{string: name of parent node (if non-root)}

\item{distance}{numeric: length of branch leading from node to this new node (if non-root). Defaults to \code{1}}

\item{count}{integer: count of cells in dataset associated with the added node}
}
\description{
Adds a node to a \code{Taxonomy} object if possible.
}
\details{
To change length of a specific branch, use \code{SetDistanceToChild}.
To change length of branches leading from a parent node, use \code{SetDistancesFromParent}.
}
