% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{ARI}
\alias{ARI}
\title{Compute adjusted Rand index}
\usage{
ARI(pred, true)
}
\arguments{
\item{pred}{numeric or character vector of predicted clusters/populations}

\item{true}{numeric or character vector of predicted clusters/populations}
}
\description{
Compute adjusted Rand index
}
