% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy.R
\name{CalibratedTaxonomy}
\alias{CalibratedTaxonomy}
\title{Build a calibrated taxonomic tree}
\usage{
CalibratedTaxonomy(name, paths, counts = NULL, round_digits = 3)
}
\arguments{
\item{name}{string: name of taxonomy}

\item{paths}{string vector: path to each node to be added}

\item{counts}{integer vector or \code{NULL}: cell count associated with each node to be added. Defaults to \code{NULL}}

\item{round_digits}{integer: how many decimal places to round weight values to, or \code{NULL}. Defaults to \code{3}}
}
\description{
Constructs a \code{Taxonomy} object with a hierarchy of labels.
Requires a vector of paths to each class, eg. \code{c'Lymphocyte', 'Lymphocyte/T Cell', 'Lymphocyte/B Cell', 'Lymphocyte/T Cell/CD4+', 'Lymphocyte/T Cell/CD8+'}.
}
\details{
Counts associated with each node can also be given, in which case weights of branches leading to each node are set as inversely proportional to the number of cells associated with it.
(This is proposed for large datasets where population sizes are considered representative and indicate how rare or common each population typically is.)

Otherwise, we set weights such that the sum of weights of branches coming out of each parent node is \code{1} and the weights are distributed uniformly among the branches.

Use \code{print} on a \code{Taxonomy} object to print a read-out of paths in the tree and its properties.
Use \code{plot} on a \code{Taxonomy} object to plot the tree structure.
}
