% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{hAccPerMatch}
\alias{hAccPerMatch}
\title{Compute Hierarchy-aware Accuracy per each ground-truth label}
\usage{
hAccPerMatch(tax, pred, true, scale = FALSE)
}
\arguments{
\item{tax}{object of class \code{Taxonomy} with computed distances}

\item{pred}{character vector of predicted populations}

\item{true}{character vector of predicted populations}
}
\description{
Computes accuracy of classification per each cell population contained in the ground-truth vector of labels, using a hierarchy-aware scoring scheme.
}
