% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{hARI}
\alias{hARI}
\title{Compute Hierarchy-aware adjusted Rand index (hARI)}
\usage{
hARI(
  tax,
  pred,
  true,
  seed = 42,
  max_npt = 1000,
  approx_batch = 500,
  approx_niter = 5000
)
}
\arguments{
\item{tax}{object of class \code{Taxonomy}}

\item{pred}{vector of predicted node labels from \code{Taxonomy}}

\item{true}{vector of true node labels from \code{Taxonomy}}

\item{seed}{integer: random seed for sampling random points to calculate \code{ExpectedIndex}}

\item{max_npt}{integer: maximum number of points for which to compute an exact solution. Default value is \code{1000}}

\item{approx_batch}{integer: size of random subset per run for approximate computation of the index value. Default value is \code{500}}

\item{approx_niter}{integer: number of random subsets (runs) to be taken for approximate computation of the index value. Default value is \code{5000}}
}
\description{
Computes the adjusted version of the Hierarchy-aware Rand index (hRI).
The adjusted index is computed as \code{AdjustedIndex = (Index - ExpectedIndex)/(MaxIndex - ExpectedIndex)}.
}
\details{
The \code{ExpectedIndex} is calculated as hARI for a simulated random clustering.
The \code{MaxIndex} is \code{1}.

Since computation of hierarchical (A)RI is expensive, exact solution is computed for cluster/label vectors of up to \code{max_npt} points.
If the number of points is higher, an average value for multiple runs on random subsets of the vectors is computed.
Size of the subset is then determined by the parameter \code{approx_batch} and the number of runs is given by \code{approx_niter}.
}
