% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{hF1PerMatch}
\alias{hF1PerMatch}
\title{Compute Hierarchy-aware F1 score per each ground-truth label}
\usage{
hF1PerMatch(tax, pred, true)
}
\arguments{
\item{tax}{object of class \code{Taxonomy} with computed distances}

\item{pred}{character vector of predicted populations}

\item{true}{character vector of predicted populations}
}
\description{
Computes the F1 score (harmonic mean of Precision and Recall) for each cell ground-truth population matched to predicted labels, using a hierarchy-aware scoring scheme.
}
