% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{hOverallAcc}
\alias{hOverallAcc}
\title{Compute Hierarchy-aware Accuracy}
\usage{
hOverallAcc(tax, pred, true)
}
\arguments{
\item{tax}{object of class \code{Taxonomy} with computed distances}

\item{pred}{numeric or character vector of predicted populations}

\item{true}{numeric or character vector of predicted populations}
}
\description{
Compute Hierarchy-aware Accuracy
}
