% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{hRI}
\alias{hRI}
\title{Compute Hierarchy-aware Rand index}
\usage{
hRI(tax, pred, true, max_npt = 1000, approx_batch = 500, approx_niter = 5000)
}
\arguments{
\item{tax}{object of class \code{Taxonomy}}

\item{pred}{numeric or character vector of predicted clusters}

\item{true}{character vector of true node labels from \code{Taxonomy}}

\item{max_npt}{integer: maximum number of points for which to compute an exact solution. Default value is \code{1000}}

\item{approx_batch}{integer: size of random subset per run for approximate computation of the index value. Default value is \code{500}}

\item{approx_niter}{integer: number of random subsets (runs) to be taken for approximate computation of the index value. Default value is \code{5000}}
}
\description{
Computes the Hierarchy-aware Rand index (hRI) based on a vector of cluster assignment, a vector of ground-truth labels and a \code{Taxonomy} object.
}
\details{
Since computation of hierarchical RI is expensive, exact solution is computed for cluster/label vectors of up to \code{max_npt} points.
If the number of points is higher, an average value for multiple runs on random subsets of the vectors is computed.
Size of the subset is then determined by the parameter \code{approx_batch} and the number of runs is given by \code{approx_niter}.
}
