% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{OverallAcc}
\alias{OverallAcc}
\title{Compute Accuracy}
\usage{
OverallAcc(pred, true)
}
\arguments{
\item{pred}{numeric or character vector of predicted populations}

\item{true}{numeric or character vector of predicted populations}
}
\description{
Compute Accuracy
}
