% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy.R
\name{Path}
\alias{Path}
\title{Get step count & distance between two nodes in a taxonomic tree}
\usage{
Path(tax, from, to, only_total = FALSE, weighted = TRUE, exp_decay = FALSE)
}
\arguments{
\item{tax}{\code{Taxonomy} object}

\item{from}{string: uniquely matchable substring identifying start node}

\item{to}{string: uniquely matchable substring identifying end node}

\item{only_total}{logical: whether to only return a single number (total distance), as opposed to rootward and leafward distances. Defaults to \code{FALSE}}

\item{weighted}{logical: whether to include weighted distances as well (or only the weighted distance, if \code{only_total} is \code{TRUE}. Defaults to \code{TRUE}}
}
\value{
If \code{only_total} is \code{TRUE} a single number (total distance, weighted if \code{weighted} is \code{TRUE}) is returned, otherwise a vector of \code{rootward} unweighted distance and \code{leafward} unweighted distance is returned (if \code{weighted} is \code{TRUE}, then weighted distances \code{rootward_w} and \code{leafward_w} are also returned).
}
\description{
Finds the shortest path between two nodes in a \code{Taxonomy} object.
}
