% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy.R
\name{PathToRoot}
\alias{PathToRoot}
\title{Get step count & distance from node to root}
\usage{
PathToRoot(tax, node, single = FALSE, weighted = TRUE)
}
\arguments{
\item{tax}{\code{Taxonomy} object}

\item{node}{string: uniquely matchable substring identifying start node}

\item{single}{logical: whether to only return a single number (total distance), as opposed to rootward and leafward distances. Defaults to \code{FALSE}}

\item{weighted}{logical: whether to include weighted distance (or only the weighted distance, if \code{single} is \code{TRUE}. Defaults to \code{TRUE}}
}
\value{
If \code{single} is \code{TRUE} a single number (total distance, weighted if \code{weighted} is \code{TRUE}) is returned, otherwise a vector of \code{unweighted} and \code{weighted} distance is returned.
}
\description{
Computes number of steps and total distance from node to root in a \code{Taxonomy} object.
}
