% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{RI}
\alias{RI}
\title{Compute Rand index}
\usage{
RI(pred, true)
}
\arguments{
\item{pred}{numeric or character vector of predicted clusters/populations}

\item{true}{numeric or character vector of predicted clusters/populations}
}
\description{
Compute Rand index
}
