% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy.R
\name{SamplingProbabilities}
\alias{SamplingProbabilities}
\title{Generate a named vector of sampling probabilites}
\usage{
SamplingProbabilities(tax)
}
\arguments{
\item{tax}{\code{Taxonomy} object}
}
\description{
Uses a \code{Taxonomy} object with known counts of cells associated with nodes in the taxonomy and produces a vector of probabilites for sampling a cell by node of taxonomy.
This is to generate a random sampling as baseline to be compared with clustering.
}
