% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy.R
\name{SetDistancesFromParent}
\alias{SetDistancesFromParent}
\title{Set length of branches leading from a parent node}
\usage{
SetDistancesFromParent(tax, name, distance)
}
\arguments{
\item{tax}{object of class \code{Taxonomy}}

\item{name}{string: path (or uniquely matchable substring thereof) of a parent node}

\item{distance}{numeric: branch length}
}
\description{
Set length of branches leading from a parent node
}
