% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy.R
\name{SetDistanceToChild}
\alias{SetDistanceToChild}
\title{Set length of a branch}
\usage{
SetDistanceToChild(tax, name, distance)
}
\arguments{
\item{tax}{object of class \code{Taxonomy}}

\item{name}{string: path (or uniquely matchable substring thereof) of a node}

\item{distance}{numeric: branch length}
}
\description{
Set length of a branch
}
