% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy.R
\name{Taxonomy}
\alias{Taxonomy}
\title{Instantiate a taxonomic tree}
\usage{
Taxonomy(name)
}
\arguments{
\item{name}{string: name of taxonomy}
}
\description{
Constructor for a \code{Taxonomy} object.
Call \code{AddNode} on a \code{Taxonomy} object to build it up.
Alternatively, use function \code{TaxonomyFromPaths} to construct the tree in one call.
}
\details{
To change length of branches, use \code{SetDistanceToChild} or \code{SetDistancesFromParent}.

Use \code{print} to print a read-out of paths in the tree and its properties.
Use \code{plot} to plot the tree structure.
}
