% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy.R
\name{TaxonomyFromPaths}
\alias{TaxonomyFromPaths}
\title{Build a taxonomic tree from node paths}
\usage{
TaxonomyFromPaths(name, paths)
}
\arguments{
\item{name}{string: name of taxonomy}

\item{paths}{string vector: path to each node to be added}
}
\description{
Constructs a \code{Taxonomy} object with a hierarchy of labels.
Requires a vector of paths to each class, eg. \code{c'Lymphocyte', 'Lymphocyte/T Cell', 'Lymphocyte/B Cell', 'Lymphocyte/T Cell/CD4+', 'Lymphocyte/T Cell/CD8+'}.
Assumes default branch lengths (all \code{1}), which can then be changed using \code{SetDistanceToChild} or \code{SetDistancesFromParent}.
}
\details{
Use \code{print} on a \code{Taxonomy} object to print a read-out of paths in the tree and its properties.
Use \code{plot} on a \code{Taxonomy} object to plot the tree structure.
}
